🐛 don't use decimals for the h value of the hsl colors
The postcss-color plugin can't handle this. We actually want to get rid of this plugin but let's fix this anyway to to avoid having issues with it. Most colors stay the same anyway, and the ones that change are only slightly different that it's not even visible to the eye.